feat: Introduce useMapPageState hook for improved state management in MapPage component
